added README.md, and additional doc fixes to illustrate roxygen bug with S4 method doc
latest version of roxygen2 is needed to effectively document lots of
other aspects of more complicated packages, but currently adds unwanted
quotation marks around aliases named by the S4 method convention. Bug
just reported to klutometis/roxygen, and so this update is posted in
support of that bug report as a reproducible example.